Lower NN precision for more speed

+++ src/bm/nnue.rs
@@ -59,11 +59,11 @@ impl Nnue {
         bytes = &bytes[INPUT * MID * 2..];
         let incremental_bias = include::bias_from_bytes_i16::<i16, MID>(bytes);
         bytes = &bytes[MID * 2..];
-        let out = Arc::new(*include::dense_from_bytes_i8::<i8, { MID * 2 }, OUTPUT>(
+        let out = Arc::new(*include::dense_from_bytes_i8::<i16, { MID * 2 }, OUTPUT>(
             bytes,
         ));
         bytes = &bytes[MID * OUTPUT * 2..];
-        let out_bias = include::bias_from_bytes_i16::<i32, OUTPUT>(bytes);
+        let out_bias = include::bias_from_bytes_i16::<i16, OUTPUT>(bytes);
         bytes = &bytes[OUTPUT * 2..];
         assert!(bytes.is_empty(), "{}", bytes.len());
